댓글

+++ src/main/java/mjucapstone/wiseculture/board/BoardService.java
@@ -6,6 +6,7 @@ import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
 import lombok.RequiredArgsConstructor;
+import mjucapstone.wiseculture.comment.CommentRepository;
 import mjucapstone.wiseculture.location.Location;
 import mjucapstone.wiseculture.member.domain.Member;
 
@@ -14,6 +15,7 @@ import mjucapstone.wiseculture.member.domain.Member;
 @RequiredArgsConstructor
 public class BoardService {
 
+	private final CommentRepository commentRepository;
 	private final BoardRepository boardRepository;
 	
 	// 게시글 목록 조회
@@ -55,23 +57,35 @@ public class BoardService {
 		if(boardRepository.findWriterById(id)
 				.orElseThrow(() -> new BoardException("작성자가 존재하지 않음"))
 				.getId() != writer.getId()) 
-			throw new BoardException("현재 회원이 작성한 게시글이 아님");
+			throw new BoardException("다른 회원의 게시글");
 			
 		// 게시글 업데이트
 		boardRepository.updateBoard(id, title, content, location);
 			
-		return boardRepository.getById(id); // 위에서 이미 존재하지 않는 경우 체크 됨
+		return boardRepository.findById(id).get(); // 위에서 이미 존재하지 않는 경우 체크 됨
 	}
 	
 	// 게시글 삭제
 	@Transactional
 	public void deletePost(Long id, Member writer) {
+		/*
 		if(boardRepository.existsById(id) == false) throw new BoardException("존재하지 않는 게시글");
 		if(boardRepository.findWriterById(id)
 				.orElseThrow(() -> new BoardException("작성자가 존재하지 않음"))
 				.getId() != writer.getId()) 
-			throw new BoardException("현재 회원이 작성한 게시글이 아님");
+			throw new BoardException("다른 회원의 게시글");
 		
+		// 댓글 삭제
+		commentRepository.deleteAllByBoard(boardRepository.findById(id).orElseThrow(() -> new BoardException("존재하지 않는 게시글")));
+		*/
+		
+		Board board = boardRepository.findById(id).orElseThrow(() -> new BoardException("존재하지 않는 게시글"));
+		if(board.getMember().getId() != writer.getId()) throw new BoardException("다른 회원의 게시글");
+		
+		// 댓글 삭제
+		commentRepository.deleteAllByBoard(board);
+		
+		// 게시글 삭제
 		boardRepository.deleteById(id);
 	}
